requirements.txt
fastapi>=0.104.0,<0.106.0
uvicorn>=0.24.0,<0.26.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
groq>=0.3.0,<0.4.0